LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/container-queries/anchored-child-transition-expected.txt


PASS Initial scale is 1
FAIL Start transition of scale property on anchored child assert_equals: expected "1.5" but got "2"
